# [SYSTEMML-540] Avoid redundant computation of cudnnPoolingForward in max_pool_backward