@mboehm7 2017-10-27
this is very nice - but please leave `nn/layers/dropout.dml` as it is - this can very easily be done automatically. After IPA, constant folding and branch removal this function already boils down to a single statement - the only reason why it makes a difference right is that inlining happens before IPA. I'll add another IPA pass that inlines such single statement functions.